+++ tests/testthat/test-weight_allom.R
@@ -9,7 +9,7 @@ test_that("weight_allom returns a vector of the same length as the equation tabl
               coords = c(-78, 40),
               new_eqtable = newtab
             )
-            expect_type(weight_test, "numeric")
+            expect_type(weight_test, "double")
             expect_equal(length(weight_test), nrow(newtab))
             expect_true(all(names(weight_test) %in% newtab$equation_id))
           })
